requirements.txt
pandas==1.*
oceania-query-fasta>=0.*
openpyxl>=3.*
jupyter>=1.0.0